Add a "Folders" Expression
This was made by @overjt, and he gave me permission to add it to my
fork. It adds a "Folders" expression, which works like the "files" one,
except that, as the name siggests, it lists all folders inside the set
route. This can be very useful for detecting specific directories for
template worlds, for example.